+++ event-store-management/src/main/java/uk/gov/justice/services/eventstore/management/shuttering/process/CommandHandlerQueueSupplierFactory.java
@@ -1,0 +1,21 @@
+package uk.gov.justice.services.eventstore.management.shuttering.process;
+
+import uk.gov.justice.services.messaging.jms.JmsCommandHandlerDestinationNameProvider;
+import uk.gov.justice.services.messaging.jms.JmsQueueBrowser;
+
+import java.util.function.BooleanSupplier;
+
+import javax.inject.Inject;
+
+public class CommandHandlerQueueSupplierFactory {
+
+    @Inject
+    private JmsQueueBrowser jmsQueueBrowser;
+
+    @Inject
+    private JmsCommandHandlerDestinationNameProvider jmsCommandHandlerDestinationNameProvider;
+
+    public BooleanSupplier isCommandHandlerQueueEmpty() {
+        return () -> jmsQueueBrowser.sizeOf(jmsCommandHandlerDestinationNameProvider.destinationName()) < 1;
+    }
+}